Update Structure of repo and update some functionality
* Move .bashrc into a new folder called .bash. Then remove some functionality of .bashrc
and put it into .bash_prompt, .bash_aliases so that the .bashrc file is not as bloated.
*  Move .inputrc into a new folder called .bash. Update its functionality to include more
keyboard shortcuts.
* Move .zshrc into a new folder called .zsh. However no functionality changes were made
* Update README with more sources used for the repo
* Update utils.sh to have the sudo prefix for some commands to facilitate symlinking and
downloading around the /opt/ and /usr/ folders. Also update the
remove_symlinks_of_source_file function to have the ability to specify a search path instead
of only using $HOME.
* Update reset.sh to remove symlinks from the new symlink locations. Also update it to
remove the new idea run script symlink.
* Update setup.sh to use new locations of dotfiles. Also add install and download code for
nvm and intellj idea scripts. Also add a symlink for intellj idea to the usr/local/bin directory
so it gets picked up in the $PATH.